requirements.txt
beautifulsoup4~=4.12.2
Requests~=2.31.0
selenium~=4.0.0
tldextract~=2.2.3
openpyxl~=3.1.2
PyYAML~=6.0.1